fix: bugs & refactor code for new conditions data structure - backport
