+++ desk/src/components/Settings/Assignment Rules/AssignmentRulesSection.vue
@@ -1,20 +1,16 @@
 <template>
-  <AssignmentConditions
+  <CFConditions
     v-if="props.conditions.length > 0"
     :conditions="props.conditions"
     :level="0"
-    :errors="props.errors"
+    :doctype="'HD Ticket'"
+    :disableAddCondition="props.errors !== ''"
   />
   <div
     v-if="props.conditions.length == 0"
     class="flex p-4 items-center cursor-pointer justify-center gap-2 text-sm border border-gray-300 text-gray-600 rounded-md"
     @click="
-      props.conditions.push({
-        field: null,
-        operator: 'equals',
-        value: '',
-        conjunction: 'and',
-      });
+      props.conditions.push(['', '', '']);
       validateAssignmentRule(props.name);
     "
   >
@@ -43,17 +39,25 @@
 <script setup lang="ts">
 import { Button, Dropdown, FeatherIcon } from "frappe-ui";
 import { watchDebounced } from "@vueuse/core";
-import {
-  validateAssignmentRule,
-  validateConditions,
-} from "../../../stores/assignmentRules";
-import AssignmentConditions from "./Assignment Conditions/AssignmentConditions.vue";
+import { validateAssignmentRule } from "../../../stores/assignmentRules";
+import CFConditions from "@/components/conditions-filter/CFConditions.vue";
+import { validateConditions } from "@/utils";
 
 const props = defineProps({
   conditions: Array<any>,
   name: String,
   errors: String,
 });
+
+const getConjunction = () => {
+  let conjunction = "and";
+  props.conditions.forEach((condition) => {
+    if (typeof condition == "string") {
+      conjunction = condition;
+    }
+  });
+  return conjunction;
+};
 
 const dropdownOptions = [
   {
@@ -65,21 +69,8 @@ const dropdownOptions = [
   {
     label: "Add condition group",
     onClick: () => {
-      const conjunction =
-        props.conditions.length > 1 ? props.conditions[1]?.conjunction : "and";
-      props.conditions.push({
-        field: "group",
-        operator: "equals",
-        value: [
-          {
-            field: null,
-            operator: "equals",
-            value: "",
-            conjunction: "and",
-          },
-        ],
-        conjunction: conjunction,
-      });
+      const conjunction = getConjunction();
+      props.conditions.push(conjunction, [[]]);
     },
   },
 ];
@@ -90,15 +81,9 @@ const addCondition = () => {
   if (!isValid) {
     return;
   }
-  const conjunction =
-    props.conditions.length > 1 ? props.conditions[1]?.conjunction : "and";
+  const conjunction = getConjunction();
 
-  props.conditions.push({
-    field: null,
-    operator: "equals",
-    value: "",
-    conjunction: conjunction,
-  });
+  props.conditions.push(conjunction, ["", "", ""]);
 };
 
 watchDebounced(
