+++ desk/src/components/Settings/Sla/utils.ts
@@ -1,24 +1,3 @@
-import { createResource } from "frappe-ui";
-
-export const filterableFields = createResource({
-  url: "helpdesk.api.doc.get_filterable_fields",
-  params: {
-    doctype: "HD Ticket",
-  },
-  transform: (data) => {
-    data = data
-      .filter((field) => !field.fieldname.startsWith("_"))
-      .map((field) => {
-        return {
-          label: field.label,
-          value: field.fieldname,
-          ...field,
-        };
-      });
-    return data;
-  },
-});
-
 export function formatTimeHMS(seconds) {
   const days = Math.floor(seconds / (3600 * 24));
   const hours = Math.floor((seconds % (3600 * 24)) / 3600);
